Move directory of homepage to be root
