+++ frontend/src/app/(auth)/leetcode-dashboard/page.tsx
@@ -1,9 +1,8 @@
 "use client";
 
 import { useAuth } from "@/components/auth/AuthContext";
-import LandingPage from "./components/landing-page/LandingPage";
-// import Dashboard from "./components/dashboard/Dashboard";
-import LeetcodeDashboard from "./leetcode-dashboard/LeetcodeDashboard";
+import LandingPage from "@/app/(home)/components/landing-page/LandingPage";
+import LeetcodeDashboard from "./LeetcodeDashboard";
 
 const Home = () => {
   const { token } = useAuth();
