+++ frontend/src/app/(home)/components/navbar/Navbar.tsx
