+++ frontend/src/app/(home)/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import Navbar from "@/app/home/components/navbar/Navbar";
-import LandingPage from "@/app/home/components/landing-page/LandingPage";
+import Navbar from "@/app/(home)/components/navbar/Navbar";
+import LandingPage from "@/app/(home)/components/landing-page/LandingPage";
 
 const Home = () => {
     return (
